+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/FirDiagnosticTestGenerated.java
@@ -160,6 +160,12 @@ public class FirDiagnosticTestGenerated extends AbstractFirDiagnosticTest {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("compiler-tests/src/test/data/diagnostic/fir/inject"), Pattern.compile("^(.+)\\.kt$"), null, true);
     }
 
+    @Test
+    @TestMetadata("SuggestInjectClassOnSingleContructor.kt")
+    public void testSuggestInjectClassOnSingleContructor() {
+      runTest("compiler-tests/src/test/data/diagnostic/fir/inject/SuggestInjectClassOnSingleContructor.kt");
+    }
+
     @Nested
     @TestMetadata("compiler-tests/src/test/data/diagnostic/fir/inject/assisted")
     @TestDataPath("$PROJECT_ROOT")
@@ -189,6 +195,32 @@ public class FirDiagnosticTestGenerated extends AbstractFirDiagnosticTest {
       @TestMetadata("FunctionMemberInjectsCannotHaveTypeParams.kt")
       public void testFunctionMemberInjectsCannotHaveTypeParams() {
         runTest("compiler-tests/src/test/data/diagnostic/fir/inject/member/FunctionMemberInjectsCannotHaveTypeParams.kt");
+      }
+    }
+  }
+
+  @Nested
+  @TestMetadata("compiler-tests/src/test/data/diagnostic/fir/interop")
+  @TestDataPath("$PROJECT_ROOT")
+  public class Interop {
+    @Test
+    public void testAllFilesPresentInInterop() {
+      KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("compiler-tests/src/test/data/diagnostic/fir/interop"), Pattern.compile("^(.+)\\.kt$"), null, true);
+    }
+
+    @Nested
+    @TestMetadata("compiler-tests/src/test/data/diagnostic/fir/interop/dagger")
+    @TestDataPath("$PROJECT_ROOT")
+    public class Dagger {
+      @Test
+      public void testAllFilesPresentInDagger() {
+        KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("compiler-tests/src/test/data/diagnostic/fir/interop/dagger"), Pattern.compile("^(.+)\\.kt$"), null, true);
+      }
+
+      @Test
+      @TestMetadata("DoNotSuggestMovingInjectAnnotationToClassWhenUsingJavaxOrJakartaInject.kt")
+      public void testDoNotSuggestMovingInjectAnnotationToClassWhenUsingJavaxOrJakartaInject() {
+        runTest("compiler-tests/src/test/data/diagnostic/fir/interop/dagger/DoNotSuggestMovingInjectAnnotationToClassWhenUsingJavaxOrJakartaInject.kt");
       }
     }
   }