+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrGraphGenerator.kt
@@ -136,12 +136,19 @@ internal class IrGraphGenerator(
     fieldInitializers += (this to init)
   }
 
+  fun IrField.initFinal(body: IrBuilderWithScope.() -> IrExpression): IrField = apply {
+    isFinal = true
+    initializer = createIrBuilder(symbol).run { irExprBody(body()) }
+  }
+
   fun generate() =
     with(graphClass) {
       val ctor = primaryConstructor!!
 
       val extraConstructorStatements =
         mutableListOf<IrBuilderWithScope.(thisReceiver: IrValueParameter) -> IrStatement>()
+
+      val thisReceiverParameter = thisReceiverOrFail
 
       fun addBoundInstanceField(
         typeKey: IrTypeKey,
@@ -161,8 +168,8 @@ internal class IrGraphGenerator(
               fieldType = symbols.metroProvider.typeWith(typeKey.type),
               fieldVisibility = DescriptorVisibilities.PRIVATE,
             )
-            .withInit(typeKey) { thisReceiver, typeKey ->
-              instanceFactory(typeKey.type, initializer(thisReceiver, typeKey))
+            .initFinal {
+              instanceFactory(typeKey.type, initializer(thisReceiverParameter, typeKey))
             }
       }
 
@@ -249,14 +256,11 @@ internal class IrGraphGenerator(
           }
         }
 
-      val thisReceiverParameter = thisReceiverOrFail
-
       // Don't add it if it's not used
       if (node.typeKey in sealResult.reachableKeys) {
         val thisGraphField =
           addSimpleInstanceField(fieldNameAllocator.newName("thisGraphInstance"), node.typeKey) {
-            thisReceiver ->
-            irGet(thisReceiver)
+            irGet(thisReceiverParameter)
           }
 
         instanceFields[node.typeKey] = thisGraphField
@@ -273,8 +277,11 @@ internal class IrGraphGenerator(
               fieldType = symbols.metroProvider.typeWith(node.typeKey.type),
               fieldVisibility = DescriptorVisibilities.PRIVATE,
             )
-            .withInit(node.typeKey) { thisReceiver, _ ->
-              instanceFactory(node.typeKey.type, irGetField(irGet(thisReceiver), thisGraphField))
+            .initFinal {
+              instanceFactory(
+                node.typeKey.type,
+                irGetField(irGet(thisReceiverParameter), thisGraphField),
+              )
             }
       }
 
@@ -377,7 +384,7 @@ internal class IrGraphGenerator(
                   fieldType = symbols.metroProvider.typeWith(node.typeKey.type),
                   fieldVisibility = DescriptorVisibilities.PRIVATE,
                 )
-                .withInit(key) { thisReceiver, fieldIndex ->
+                .withInit(key) { thisReceiver, _ ->
                   // If this is in instance fields, just do a quick assignment
                   if (binding.typeKey in instanceFields) {
                     val field = instanceFields.getValue(binding.typeKey)
@@ -506,7 +513,10 @@ internal class IrGraphGenerator(
 
       val finalConstructorStatements:
         List<IrBuilderWithScope.(thisReceiver: IrValueParameter) -> IrStatement>
-      if (fieldInitializers.size + extraConstructorStatements.size > STATEMENTS_PER_METHOD) {
+      if (
+        options.chunkFieldInits &&
+          fieldInitializers.size + extraConstructorStatements.size > STATEMENTS_PER_METHOD
+      ) {
         // Larger graph, split statements
         // Chunk our constructor statements and split across multiple init functions
         val chunks =
@@ -552,12 +562,10 @@ internal class IrGraphGenerator(
         // Small graph, just do it in the constructor
         // Assign those initializers directly to their fields and mark them as final
         for ((field, init) in fieldInitializers) {
-          field.isFinal = true
-          field.initializer =
-            createIrBuilder(field.symbol).run {
-              val typeKey = fieldsToTypeKeys.getValue(field)
-              irExprBody(init(thisReceiverParameter, typeKey))
-            }
+          field.initFinal {
+            val typeKey = fieldsToTypeKeys.getValue(field)
+            init(thisReceiverParameter, typeKey)
+          }
         }
         finalConstructorStatements = extraConstructorStatements
       }
@@ -565,7 +573,7 @@ internal class IrGraphGenerator(
       // Add extra constructor statements
       with(ctor) {
         val originalBody = checkNotNull(body)
-        buildBlockBody() {
+        buildBlockBody {
           +originalBody.statements
           for (statement in finalConstructorStatements) {
             +statement(thisReceiverParameter)
@@ -692,14 +700,14 @@ internal class IrGraphGenerator(
   private fun IrClass.addSimpleInstanceField(
     name: String,
     typeKey: IrTypeKey,
-    initializerExpression: IrBuilderWithScope.(thisReceiver: IrValueParameter) -> IrExpression,
+    initializerExpression: IrBuilderWithScope.() -> IrExpression,
   ): IrField =
     addField(
         fieldName = name,
         fieldType = typeKey.type,
         fieldVisibility = DescriptorVisibilities.PRIVATE,
       )
-      .withInit(typeKey) { thisReceiver, _ -> initializerExpression(thisReceiver) }
+      .initFinal { initializerExpression() }
 
   private fun DependencyGraphNode.implementOverrides(
     context: GraphGenerationContext,