+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroGradleSubplugin.kt
@@ -64,7 +64,7 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
   ): Provider<List<SubpluginOption>> {
     val project = kotlinCompilation.target.project
     val extension = project.extensions.getByType(MetroPluginExtension::class.java)
-    val platformCanGnerateContributionHints =
+    val platformCanGenerateContributionHints =
       when (kotlinCompilation.platformType) {
         KotlinPlatformType.common,
         KotlinPlatformType.jvm,
@@ -112,7 +112,7 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
         add(
           lazyOption(
             "generate-contribution-hints",
-            extension.generateContributionHints.orElse(platformCanGnerateContributionHints),
+            extension.generateContributionHints.orElse(platformCanGenerateContributionHints),
           )
         )
         add(
@@ -124,7 +124,14 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
         add(lazyOption("enable-scoped-inject-class-hints", extension.enableScopedInjectClassHints))
         add(lazyOption("transform-providers-to-private", extension.transformProvidersToPrivate))
         add(lazyOption("shrink-unused-bindings", extension.shrinkUnusedBindings))
+        add(lazyOption("chunk-field-inits", extension.chunkFieldInits))
         add(lazyOption("public-provider-severity", extension.publicProviderSeverity))
+        add(
+          lazyOption(
+            "warn-on-inject-annotation-placement",
+            extension.warnOnInjectAnnotationPlacement,
+          )
+        )
         add(
           lazyOption(
             "enable-top-level-function-injection",